Elements: Move regtest classes to its own package
