+++ wallets/elementsd/src/test/java/bisq/wallets/elementsd/ElementsdSendAndListTxsIntegrationTests.java
@@ -18,6 +18,7 @@
 package bisq.wallets.elementsd;
 
 import bisq.wallets.core.model.AddressType;
+import bisq.wallets.elementsd.regtest.ElementsdRegtestSetup;
 import bisq.wallets.elementsd.rpc.responses.ElementsdListTransactionsResponseEntry;
 import org.junit.jupiter.api.Test;
